Django Video Chat Application/DjangoGirlsWorkShopNov24/requirements.txt
asgiref
async-timeout
attrs
autobahn
Automat
certifi
cffi
channels
charset-normalizer
constantly
cryptography
daphne
Django
django-braces
django-browser-reload
django-debug-toolbar
django-embed-video
djangorestframework
hyperlink
idna
incremental
msgpack
pillow>=11.0.0
pyasn1
pyasn1_modules
pycparser
pymemcache
pyOpenSSL
requests
service-identity
sqlparse
Twisted
txaio
typing_extensions
urllib3
zope.interface
